Remove fs_file_open_handle().
fs_file_open_handle() was a hack to support stdin/stdout/stderr. With
this commit a set of new variables of type `const char*` are exposed
which can be passed into fs_file_open():

  - FS_STDIN
  - FS_STDOUT
  - FS_STDERR

When a file path pointer equal to one of these variables is passed into
fs_file_open(), the library will detect it and pass it straight to the
backend. The backend will then detect it for itself and handle it.

To open a handle to stdout, one can do this:

    fs_file_open(NULL, FS_STDOUT, FS_WRITE, &pStdOut);

The file must still be closed with fs_file_close().