Rework fs_mkdir().
This now returns FS_ALREADY_EXISTS if the directory already exists.

If FS_NO_CREATE_DIRS is specified in the options, creation of the full
directory structure will not be attempted, in which case the result code
FS_DOES_NOT_EXIST will be returned if a parent directory does not exist.