Do a full garbage collection on unmounting an archive.
The reason for this change is that it most likely means the application
is done with that archive.

Currently this does a full GC which means it'll collect other
unreferenced archives. A better solution to this would be to rework the
GC system to allow collection of only a specific archive.